requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.26
python-dotenv>=1.0
pydantic>=2.6

# LangChain + OpenAI
langchain>=0.2
langchain-core>=0.2
langchain-openai>=0.1
